+++ packages/picasso-codemod/src/v20.0.0/remove-tooltip-variant/__testfixtures__/basic.input.tsx
@@ -1,0 +1,14 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
+
+import { Tooltip } from '@toptal/picasso'
+
+const Example = () => (
+  <div>
+    <Tooltip content='Content goes here...' open variant='dark'>
+      <Button>Test me</Button>
+    </Tooltip>
+  </div>
+)
+
+export default Example