+++ packages/picasso/src/TypographyOverflow/story/TooltipVariants.example.tsx
@@ -4,16 +4,10 @@ import { TypographyOverflow } from '@toptal/picasso'
 const Example = () => {
   return (
     <div style={{ width: 300, marginTop: 100 }}>
-      <TypographyOverflow
-        data-testid='ellipsed-text-dark'
-        tooltipVariant='dark'
-      >
+      <TypographyOverflow data-testid='ellipsed-text-dark'>
         Dark tooltip. This typography is very long and therefore it overflows.
       </TypographyOverflow>
-      <TypographyOverflow
-        data-testid='ellipsed-text-light'
-        tooltipVariant='light'
-      >
+      <TypographyOverflow data-testid='ellipsed-text-light'>
         Light tooltip. This typography is very long and therefore it overflows.
       </TypographyOverflow>
     </div>